functional node windows

+++ platform/node/build.js
@@ -1,36 +1,37 @@
 import esbuild from "esbuild";
-import child_process from "node:child_process";
-import os from "node:os";
 import path from "node:path";
 import url from "node:url";
 import fs from "node:fs";
 import { buildNodeBinding } from "./build-binding.js";
 
 const currentDirectory = path.dirname(url.fileURLToPath(import.meta.url));
-const rootDirectory = path.resolve(currentDirectory, "..", "..");
 
-esbuild.buildSync({
-    entryPoints: ["src/index.ts"],
-    outfile: "index.js",
-    bundle: true,
-    format: "esm",
-    packages: "external",
-    platform: "node"
-});
+if (process.argv.includes("--binding")) {
+    buildNodeBinding(currentDirectory);
+} else {
+    const rootDirectory = path.resolve(currentDirectory, "..", "..");
 
-const fullstackedModulesDirectory = path.resolve(
-    currentDirectory,
-    "fullstacked_modules"
-);
+    esbuild.buildSync({
+        entryPoints: ["src/index.ts"],
+        outfile: "index.js",
+        bundle: true,
+        format: "esm",
+        packages: "external",
+        platform: "node"
+    });
 
-if (fs.existsSync(fullstackedModulesDirectory)) {
-    fs.rmSync(fullstackedModulesDirectory, { recursive: true });
+    const fullstackedModulesDirectory = path.resolve(
+        currentDirectory,
+        "fullstacked_modules"
+    );
+
+    if (fs.existsSync(fullstackedModulesDirectory)) {
+        fs.rmSync(fullstackedModulesDirectory, { recursive: true });
+    }
+
+    fs.cpSync(
+        path.resolve(rootDirectory, "fullstacked_modules"),
+        fullstackedModulesDirectory,
+        { recursive: true }
+    );
 }
-
-fs.cpSync(
-    path.resolve(rootDirectory, "fullstacked_modules"),
-    fullstackedModulesDirectory,
-    { recursive: true }
-);
-
-buildNodeBinding(currentDirectory);
